+++ libre_pvz/src/plant/peashooter.rs
@@ -22,8 +22,11 @@ use bevy::prelude::*;
 use bevy::sprite::Anchor;
 use bevy_asset_loader::prelude::*;
 use libre_pvz_animation::transform::SpriteBundle2D;
-use libre_pvz_resources::cached::Cached;
-use libre_pvz_resources::model::{MarkerRegistryExt, Model, ModelState, StateTransitionEvent};
+use libre_pvz_resources::cache_known_states;
+use libre_pvz_resources::model::{
+    MarkerRegistryExt, Model, ModelState, ModelSystem,
+    CoolDown, StateTransitionEvent, TransitionTrigger,
+};
 use crate::core::kinematics::{Position, Velocity};
 use crate::core::projectile::Projectile;
 use crate::scene::loading::AssetState;
@@ -34,9 +37,16 @@ pub struct PeashooterPlugin;
 
 impl Plugin for PeashooterPlugin {
     fn build(&self, app: &mut App) {
-        app.add_system_set(SystemSet::on_update(AssetState::AssetReady).with_system(peashooter_fire_system))
-            .register_marker::<Peashooter>("Peashooter")
-            .register_marker::<PeashooterHead>("PeashooterHead");
+        app.register_marker::<Peashooter>("Peashooter")
+            .register_marker::<PeashooterHead>("PeashooterHead")
+            .add_system_set(SystemSet::on_enter(AssetState::AssetReady)
+                .with_system(initialize_state_index_system))
+            .add_system_set(SystemSet::on_update(AssetState::AssetReady)
+                .with_system(peashooter_fire_system))
+            .add_system_set(SystemSet::on_update(AssetState::AssetReady)
+                .with_system(peashooter_force_shooting_system)
+                .before(ModelSystem::TransitionTrigger)
+                .after(ModelSystem::CoolDownTicking));
     }
 }
 
@@ -57,30 +67,59 @@ pub struct Peashooter;
 #[derive(Default, Debug, Copy, Clone, Component)]
 pub struct PeashooterHead;
 
-#[allow(unused)]
-static IDLE: Cached<&str, usize> = Cached::new("idle");
-static SHOOTING: Cached<&str, usize> = Cached::new("shooting");
+cache_known_states! {
+    idle,
+    shooting_1,
+    shooting_2,
+}
 
 const PEA_VELOCITY: f32 = 9.9 * 30.0;
 
+fn initialize_state_index_system(
+    assets: Res<PeashooterAssets>,
+    models: Res<Assets<Model>>,
+    mut commands: Commands,
+) {
+    let root = models.get(&assets.model).unwrap();
+    let head = root.attachments[0].child_model.get(&models).unwrap();
+    match StateIndex::cache(&head.states) {
+        Ok(indices) => commands.insert_resource(indices),
+        Err(err) => error!("state name mismatch: {err}"),
+    }
+}
+
+/// Force the Peashooters to shoot bullet peas (for debugging purposes).
+fn peashooter_force_shooting_system(
+    mut head: Query<(Entity, &mut CoolDown, &ModelState), With<PeashooterHead>>,
+    mut triggers: EventWriter<TransitionTrigger>,
+    models: Res<Assets<Model>>,
+    states: Res<StateIndex>,
+) {
+    for (entity, mut cool_down, state) in &mut head {
+        if state.current_state != states.idle { continue; }
+        if let Some(trigger) = state.trigger_if_ready(
+            entity, &models, &mut cool_down, "shoot") {
+            triggers.send(trigger);
+        }
+    }
+}
+
 /// Fire a pea every time we enter the `SHOOTING` state.
-pub fn peashooter_fire_system(
+fn peashooter_fire_system(
     head: Query<(&ModelState, &GlobalTransform), With<PeashooterHead>>,
     mut transitions: EventReader<StateTransitionEvent>,
-    models: Res<Assets<Model>>,
     assets: Res<PeashooterAssets>,
+    states: Res<StateIndex>,
     mut commands: Commands,
 ) {
     for &StateTransitionEvent { target_entity, .. } in transitions.iter() {
         let (state, trans) = head.get(target_entity).unwrap();
-        let shooting = SHOOTING.get_handle_or_lazy_init(||
-            &models.get(state.model()).unwrap().states).unwrap();
         // we want to keep that line aligned with other assignments
         #[allow(clippy::field_reassign_with_default)]
-        if state.current_state == shooting {
+        if [states.shooting_1, states.shooting_2].contains(&state.current_state) {
             let trans = trans.translation();
-            let x0 = trans.x + 24.0;
-            let y0 = trans.y + 33.0;
+            let x0 = trans.x + (20.0 + 40.0);
+            let y0 = trans.y + (30.0 - 40.0);
             let p0 = Position(Vec3::new(x0, y0, 0.0));
             let vel = Velocity(Vec3::new(PEA_VELOCITY, 0.0, 0.0));
             let mut bundle = SpriteBundle2D::default();